Bluewing: Hole punch + IPv6 fixes
Hole punch was working before, so after some testing some issues were ironed out.
The server port was not set to re-use after hole punch, so it did not inherit. The client port was not set to re-use after local port was set, so it did not inherit either. The client connection would try too hard to stay alive in error scenarios, lingering too long. The client didn't always have the server address available right after the connection attempt - albeit mostly in error scenarios.
IPv6 permanent addresses do not have an infinite valid lifetime; which makes sense as they may be router-assigned. Using prefix-length was valid, but limits to Vista+ as XP doesn't have the prefix length - but to be honest, IPv6 on XP does not mesh well with servers anyway, due to XP not permitting IPv4 + IPv6 hosting on the same server port.
IPv6 netlink alternative was non-functional due to checking if something succeeded, not failed. It also ignored IPv6 addresses due to them being under a different enum type.

IPv6 permanent address lookup was doomed to fail on Android, as Android doesn't use privacy addresses by default, although it does have multiple addresses. The flags were incorrectly checked - Windows flags are very different (and more useful).

+++ DarkEdif/Lib/Shared/Lacewing/deps/http-parser/http_parser.c
@@ -1948,7 +1948,7 @@ reexecute:
 	  case s_body_identity:
 	  {
 		uint64_t to_read = MIN(parser->content_length,
-							   (uint64_t) (((uint64_t)data + len) - (uint64_t)p));
+							   (uint64_t) ((data + len) - p));
 
 		assert(parser->content_length != 0
 			&& parser->content_length != ULLONG_MAX);
@@ -2082,7 +2082,7 @@ reexecute:
 	  case s_chunk_data:
 	  {
 		uint64_t to_read = MIN(parser->content_length,
-							   (uint64_t) (((uint64_t)data + len) - (uint64_t)p));
+							   (uint64_t) ((data + len) - p));
 
 		assert(parser->flags & F_CHUNKED);
 		assert(parser->content_length != 0
